Fix #8958: Add initialization ordering and remove user/bot initialization from Authorizer (#8959) merge in 0.12.3
