+++ openmetadata-service/src/main/java/org/openmetadata/service/OpenMetadataApplication.java
@@ -97,9 +97,7 @@ public class OpenMetadataApplication extends Application<OpenMetadataApplication
       throws ClassNotFoundException, IllegalAccessException, InstantiationException, NoSuchMethodException,
           InvocationTargetException, IOException {
     // init email Util for handling
-    if (catalogConfig.getSmtpSettings() != null && catalogConfig.getSmtpSettings().getEnableSmtpServer()) {
-      EmailUtil.EmailUtilBuilder.build(catalogConfig.getSmtpSettings());
-    }
+    EmailUtil.initialize(catalogConfig);
     final Jdbi jdbi = createAndSetupJDBI(environment, catalogConfig.getDataSourceFactory());
     final SecretsManager secretsManager =
         SecretsManagerFactory.createSecretsManager(
