+++ openmetadata-service/src/main/java/org/openmetadata/service/resources/teams/RoleResource.java
@@ -105,6 +105,7 @@ public class RoleResource extends EntityResource<Role, RoleRepository> {
               policy.setId(ref.getId());
             }
             dao.initSeedData(role);
+            RoleCache.initialize();
           } catch (Exception e) {
             LOG.warn("Failed to initialize the {} from file {}", Entity.ROLE, jsonFile, e);
           }
